[enh] Migrate to cordova@11.1.0 and cordova-android@10.1.2 [enh] Migrate to nodejs 16.17.0 [enh] Use ionic cli 6.20.9 (v7 not support Ionic v1 anymore) [fix] Remove cordova-plugin-whitelist and cordova-plugin-compat (not need anymore)

+++ scripts/build-android.sh
@@ -15,6 +15,11 @@ fi
 
 cd ${PROJECT_DIR}
 
+if [[ ! -d "${ANDROID_SDK_CLI_ROOT}/bin" ]]; then
+  echo "Failed to find the Android SDK CLI. Please run first:  \`scripts/install-android-sdk-tools.sh\`"
+  exit 1
+fi
+
 # Run the build
 echo "Cleaning previous android APK files..."
 rm -rf ${ANDROID_OUTPUT_APK_DEBUG}/*.apk
